# user-service/src/main/java/com/example/userservice/controller/RoleController.java
package com.example.userservice.controller;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/roles")
public class RoleController {
    @GetMapping
    public List<Map<String, Object>> listRoles() {
        // TODO: Implement role listing
        return new ArrayList<>();
    }

    @PostMapping
    public Map<String, Object> createRole(@RequestBody Map<String, Object> roleCreateRequest) {
        // TODO: Implement role creation
        return new HashMap<>();
    }

    @PatchMapping("/{id}")
    public Map<String, Object> updateRole(@PathVariable String id, @RequestBody Map<String, Object> roleUpdateRequest) {
        // TODO: Implement role update
        return new HashMap<>();
    }

    @DeleteMapping("/{id}")
    public void deleteRole(@PathVariable String id) {
        // TODO: Implement role deletion
    }
}