# user-service/src/main/java/com/example/userservice/controller/UserController.java
package com.example.userservice.controller;

import java.util.ArrayList;
import java.util.HashMap;
import java.util.List;
import java.util.Map;

import org.springframework.web.bind.annotation.DeleteMapping;
import org.springframework.web.bind.annotation.GetMapping;
import org.springframework.web.bind.annotation.PatchMapping;
import org.springframework.web.bind.annotation.PathVariable;
import org.springframework.web.bind.annotation.PostMapping;
import org.springframework.web.bind.annotation.RequestBody;
import org.springframework.web.bind.annotation.RequestMapping;
import org.springframework.web.bind.annotation.RequestParam;
import org.springframework.web.bind.annotation.RestController;

@RestController
@RequestMapping("/users")
public class UserController {
    @PostMapping
    public Map<String, Object> createUser(@RequestBody Map<String, Object> userCreateRequest) {
        // TODO: Implement user creation
        return new HashMap<>();
    }

    @GetMapping
    public List<Map<String, Object>> listUsers(
            @RequestParam(required = false) String filter,
            @RequestParam(required = false) String sort,
            @RequestParam(required = false) Integer page,
            @RequestParam(required = false) Integer size) {
        // TODO: Implement user listing
        return new ArrayList<>();
    }

    @GetMapping("/{id}")
    public Map<String, Object> getUser(@PathVariable String id) {
        // TODO: Implement get user by id
        return new HashMap<>();
    }

    @PatchMapping("/{id}")
    public Map<String, Object> updateUser(@PathVariable String id, @RequestBody Map<String, Object> userUpdateRequest) {
        // TODO: Implement user update
        return new HashMap<>();
    }

    @DeleteMapping("/{id}")
    public void deleteUser(@PathVariable String id) {
        // TODO: Implement user deletion
    }

    @PostMapping("/{id}/roles")
    public Map<String, Object> updateUserRoles(@PathVariable String id, @RequestBody Map<String, Object> userRoleUpdateRequest) {
        // TODO: Implement add/remove roles for user
        return new HashMap<>();
    }
}